ai_deep_research_agent/requirements.txt
google-genai
firecrawl
streamlit
firecrawl-py
python-dotenv